finish splitting up parse.rs
